refactor(debug-overlay): physically split snapshot store responsibilities
